Quiet warnings about 64-to-32-bit sized integer types in GlkClient.

+++ GlkClient/GlkMemoryStream.h
@@ -22,16 +22,16 @@
 @interface GlkMemoryStream : NSObject<GlkStream> {
 	unsigned char* memory;
 	char* type;
-	int length;
+	NSInteger length;
 
-	int pointer;
+	NSInteger pointer;
 	gidispatch_rock_t rock;
 }
 
 - (id) initWithMemory: (unsigned char*) mem					// Constructs this object with the given memory
-			   length: (int) length;
+			   length: (NSInteger) length;
 - (id) initWithMemory: (unsigned char*) mem					// Constructs this object with the given memory and registers the memory
-			   length: (int) length
+			   length: (NSInteger) length
 				 type: (char*) glkType;
 
 @end
